requirements.txt
astor==0.8.1
backoff==2.2.1
esprima==4.0.1
ipython==8.12.3
json5==0.9.25
jsonlines==4.0.0
chromadb==0.3.29
langchain_chroma==0.2.2
langchain_openai==0.3.8
nbclient==0.10.2
nbformat==5.10.4
numpy==2.2.3
openai==1.65.5
pandas==2.2.3
pydantic==2.10.6
pylatexenc==2.10
PyYAML==6.0.2
Requests==2.32.3
scipy==1.15.2
sympy==1.13.1
tenacity==8.2.3
tqdm==4.67.1
vllm==0.6.6.post1